+++ src/models/DefaultPlugin.js
@@ -91,12 +91,17 @@ class DefaultPlugin {
   }
 
   /**
-   * Draws all data in the html element defined by the id.
-   * @param {string} id - Html id, without '#'.
-   * @param {boolean} readOnly - Make the draw read-only.
+   * Init drawing context of plugin.
    */
-  draw(id, readOnly) {
-    this.__drawer.draw(id, readOnly);
+  initDrawingContext() {
+    this.__drawer.initDrawingContext();
+  }
+
+  /**
+   * Draws all data in the html element defined by the id.
+   */
+  draw() {
+    this.__drawer.draw();
   }
 
   /**